Re-work factory into a delegator and autowire it in ConfigProvider adding dev dependency on zend-servicemanager
